+++ backend_servers/responses/BuildingList.go
@@ -1,6 +1,6 @@
 package response
 
 type BuildingList struct {
-	Id   uint
+	Id   uint64
 	Name string
 }
